Highlight the TOC "comments" link in red if changed
The comment link is highlighted now in case of new comments or a new
diary post.

It works across browser restarts, but it doesn't work with different
website languages because the language dependent link text is stored and
used for comparison. If switching the language, all comment links will
be highlighted again.