Get the TOC again after openstreetmap.org internal change
The class "post_heading" was removed on 2021-11-24:
https://github.com/openstreetmap/openstreetmap-website/commit/cb68450ac7b7a8dec6d0d721b9a7314bf1c7e95f

The article links couldn't be found anymore and thus the TOC was empty
(duplication of the older/newer entries still worked).
Use the class "mb-3" that is used in the same <div> now. This is not the
most future-proof solution, but the simplest/quickest.